Refactored NN api routes into own file
